+++ scripts/generate_appcast.sh
@@ -38,4 +38,6 @@ else
         "$APPCAST_WORK_DIR"
 fi
 
+perl -0pi -e "s#https://raw.githubusercontent.com/LeonTOfficial/BatteryPanic/main/Battery\\.Panic\\.$VERSION\\.md#https://github.com/LeonTOfficial/BatteryPanic/releases/tag/v$VERSION#g" "$ROOT_DIR/appcast.xml"
+
 echo "Generated appcast: $ROOT_DIR/appcast.xml"